+++ CHANGELOG.md
@@ -27,7 +27,8 @@
         - Added `expanded` option to specify initial leaf expanded level
         - Added `collapsable` option to hide a toggle element when a leaf is expaned
         - Prevented large tree growing by keeping already visited data leafs collapsed
-    - Add signature preset on report page
+    - Added signature preset on report page
+    - Added button views: `button`, `button-primary`, `button-danger` and `button-warning`
 
 ## 1.0.0-beta.8 (30-01-2019)